Separated the code for parallel embedding and tests into independent scripts
