+++ tests/test_qubit_layout.py
@@ -67,100 +67,3 @@ class TestDrawing(unittest.TestCase):
                                                   edge_vmin=-0.1, edge_vmax=0)
         dnx.drawing.qubit_layout.draw_qubit_graph(G, pos, None, quadratic_biases,
                                                   edge_vmin=0.0, edge_vmax=10)
-
-
-    # def test_visualize_parallel_embeddings(self):
-    #     # NB, matplotlib.pyplot.show() can be added to display plots
-    #     import matplotlib.pyplot as plt
-
-    #     embeddings = [{}]
-    #     T = dnx.chimera_graph(2)
-    #     n = generate_node_color_dict(T, embeddings)
-    #     e = generate_edge_color_dict(T, embeddings, node_color_dict=n)
-    #     self.assertTrue(np.all(np.isnan(list(n.values()))))
-    #     self.assertEqual(len(e), 0)
-    #     blocks_of = [1, 8]
-    #     one_to_iterable = [True, False]
-    #     for b, o in product(blocks_of, one_to_iterable):
-    #         if o:
-    #             embeddings = [
-    #                 {0: tuple(n + idx * b for n in range(b))}
-    #                 for idx in range(T.number_of_nodes() // b)
-    #             ]  # Blocks of 8
-    #         else:
-    #             embeddings = [
-    #                 {n: n + idx * b for n in range(b)}
-    #                 for idx in range(T.number_of_nodes() // b)
-    #             ]  # Blocks of 8
-
-    #         n = generate_node_color_dict(T, embeddings, one_to_iterable=o, shuffle_colormap=False)
-    #         e = generate_edge_color_dict(T, embeddings, one_to_iterable=o)
-    #         #n, e = visualize_embeddings(
-    #         #   T, embeddings, one_to_iterable=o, shuffle_colormap=False
-    #         #)
-    #         self.assertEqual(len(e), len(embeddings) * (b // 2) * (b // 2))
-
-    #         n1 = generate_node_color_dict(T, embeddings, shuffle_colormap=True, seed=42, one_to_iterable=o)
-    #         e1 = generate_edge_color_dict(T, embeddings, one_to_iterable=o, node_color_dict=n1)
-    #         # n1, e1 = visualize_embeddings(
-    #         #     T, embeddings, shuffle_colormap=True, seed=42, one_to_iterable=o
-    #         # )
-    #         if b == 1:
-    #             # Highly unlikely to have aligned colors.
-    #             self.assertFalse(
-    #                 all([n1[idx] <= n1[idx + 1] for idx in range(len(n1) - 1)])
-    #             )
-    #             self.assertTrue(
-    #                 all([n[idx] <= n[idx + 1] for idx in range(len(n) - 1)])
-    #             )
-    #         n2 = generate_node_color_dict(T, embeddings, shuffle_colormap=True, seed=42, one_to_iterable=o)
-    #         e2 = generate_edge_color_dict(T, embeddings, one_to_iterable=o, node_color_dict=n1)
-    #         # n2, e2 = visualize_embeddings(
-    #         #     T, embeddings, shuffle_colormap=True, seed=42, one_to_iterable=o
-    #         # )
-    #         self.assertEqual(n1, n2)
-    #         self.assertEqual(e1, e2)
-
-    #     S = nx.Graph()
-    #     S.add_node(0)
-        # embs = [{0: n} for n in T.nodes]
-        # n, e = visualize_embeddings(
-        #     T, embeddings=embs, S=S
-        # )  # Should plot every node colorfully but no edges
-        # vals = np.array(list(n.values()))
-        # self.assertTrue(np.all(np.logical_and(vals < len(embs), vals >= 0)))
-        # self.assertEqual(len(e), 0)
-
-        # S.add_edges_from(list(T.edges)[:2])
-        # emb = {n: n for n in T.nodes}
-
-        # n, e = visualize_embeddings(
-        #     T, embeddings=[emb], S=S
-        # )  # Should plot 3 nodes, and two edges
-
-        # vals = np.array(list(n.values()))
-        # self.assertEqual(len(vals), T.number_of_nodes())
-        # self.assertEqual(np.sum(np.isnan(vals)), T.number_of_nodes() - 3)
-        # self.assertEqual(len(e), 2)
-
-        # n, e = visualize_embeddings(
-        #     T, embeddings=[emb], S=None
-        # )  # Should plot every nodes and edges
-        # self.assertEqual(len(e), T.number_of_edges())
-        # self.assertEqual(len(n), T.number_of_nodes())
-
-        # S = nx.from_edgelist({(i, i + 1) for i in range(2)})
-        # embedding = {i: (i, i + 4) for i in range(3)}
-        # embeddings = [
-        #     embedding,
-        #     {k: tuple(v + 8 for v in c) for k, c in embedding.items()},
-        # ]
-        # plt.figure("check one_to_iterable with S")
-        # n, e = visualize_embeddings(T, S=S, embeddings=embeddings, one_to_iterable=True)
-        # vals = np.array(list(n.values()))
-        # self.assertEqual(len(e), 7 * 2)
-        # self.assertEqual(len(vals), T.number_of_nodes())
-        # self.assertEqual(np.sum(np.isnan(vals)), T.number_of_nodes() - 6 * 2)
-        # # Should plot 7 edges total and 6 nodes: Top left and top right cells.
-        # #  3 (NE to SW) chains, 2 logical double-couplers
-        # # NB without S would be a clique.
